chore(ci): update .github/workflows/update-docs.yml to use latest actions/cache (#4407)
Hi all!

GitHub will be deprecating older versions of `actions/cache` on March
1st, including version `4.0.2` which is used in
`.github/workflows/update-docs.yml`.

This is a quick PR to get that workflow updated to the newest version.

See [GitHub's
notice](https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down)
and the [full list of deprecated
versions](https://github.com/actions/cache/discussions/1510) for more
info.